Fix the two CI failures, and close the gap that let one through
The 32-bit job caught a real hole. Three test assertions passed an untyped
constant to FromInt, which infers int -- 32 bits on a 386 or arm target, so
FromInt(math.MinInt64) does not compile there. The library itself was fine;
the build step passed and only the test step failed.

One of the three could not be fixed with a cast. It guarded a 64-bit
literal behind `if bits := ^uint(0) >> 63; bits == 1`, which does nothing:
Go type-checks both branches, so the constant still has to fit. It now
checks the platform's own boundary with math.MinInt and math.MaxInt, which
is the invariant that actually matters for Int and Uint and holds on 32-
and 64-bit alike.

`make cross` used `go build`, and build ignores _test.go files -- which is
exactly why this passed locally and failed in CI. It now uses `go vet`,
which type-checks the tests, and the CI per-architecture step does the
same. Both would have caught this before the push.

The lint job failed for an unrelated reason: golangci-lint-action v6 only
understands golangci-lint v1 version strings, so pinning v2.13.1 made it
error out. Bumped to the v7 action.

Worth recording what did pass: the s390x job ran the full suite under QEMU
and was green, so the big-endian word accessors are now verified by an
actual test run rather than only by compiling.

+++ semantics_test.go
@@ -163,9 +163,11 @@ func TestDivideByZero(t *testing.T) {
 // The integer conversions rely on wrapping arithmetic at the extremes; these are
 // the values where a sign test done the obvious way gets it wrong.
 func TestConversionExtremes(t *testing.T) {
-	assert.Equal(t, int32(math.MinInt32), FromInt(math.MinInt32).Int32(), "int32 minimum round-trips")
-	assert.Equal(t, int64(math.MinInt64), FromInt(math.MinInt64).Int64(), "int64 minimum round-trips")
-	assert.Equal(t, int32(math.MaxInt32), FromInt(math.MaxInt32).Int32(), "int32 maximum")
+	// The argument types are explicit: an untyped constant infers int, which is
+	// 32 bits on a 32-bit target, so FromInt(math.MinInt64) will not compile there.
+	assert.Equal(t, int32(math.MinInt32), FromInt(int32(math.MinInt32)).Int32(), "int32 minimum round-trips")
+	assert.Equal(t, int64(math.MinInt64), FromInt(int64(math.MinInt64)).Int64(), "int64 minimum round-trips")
+	assert.Equal(t, int32(math.MaxInt32), FromInt(int32(math.MaxInt32)).Int32(), "int32 maximum")
 	assert.Equal(t, uint64(math.MaxUint64), FromUint(uint64(math.MaxUint64)).Uint64(), "uint64 maximum")
 
 	// One past each boundary must fail.
